Rerererefactor

+++ src/main/java/dev/misei/einfachml/repository/model/NetworkBoard.java
@@ -24,6 +24,6 @@ public class NetworkBoard {
     //Current Status
     Status status;
 
-    int datasetSize;
-    int predictionsSize;
+    long datasetSize;
+    long predictionsSize;
 }
